slackbot/requirements.txt
flask==1.1.4
slackeventsapi==2.2.1
slackclient==2.9.3
